♻️ refactor(server): relocate scanner pipeline (Scanner/Analyzer/ScanCoordinator/ScanOrchestrator/DocumentCollector) to commonMain
